# server/src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "공모아 API 문서",
    "description": "공모아 서버 API관련 문서입니다.",
    "version": "1.0.0"
  },
  "host": "https://gongmoa.site/api",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/isMenuValid": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 불러올때 사용여부 확인 로직",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "menu_path": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/adminMenuList": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 리스트 불러오기 로직",
        "description": "",
        "parameters": [
          {
            "name": "pg",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pp",
            "in": "query",
            "type": "string"
          },
          {
            "name": "parentCode",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/setAdminMenuStatus": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 상태변경 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "example": "any"
                },
                "menu_code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/getAdminMenuCode": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 부모코드 조회 로직",
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/getAdminMenuDetail": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 상세 조회 로직",
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/modifyAdminMenu": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 메뉴 수정 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "act": {
                  "example": "any"
                },
                "menu_code": {
                  "example": "any"
                },
                "menu_name": {
                  "example": "any"
                },
                "menu_path": {
                  "example": "any"
                },
                "use_yn": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/setAdminMenuOrder": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 메뉴 순서변경 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "parentCode": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/admin_menu/isMenuValid": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 불러올때 사용여부 확인 로직",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "menu_path": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin_menu/adminMenuList": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 리스트 불러오기 로직",
        "description": "",
        "parameters": [
          {
            "name": "pg",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pp",
            "in": "query",
            "type": "string"
          },
          {
            "name": "parentCode",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin_menu/setAdminMenuStatus": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 상태변경 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "example": "any"
                },
                "menu_code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin_menu/getAdminMenuCode": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 부모코드 조회 로직",
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin_menu/getAdminMenuDetail": {
      "get": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 상세 조회 로직",
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin_menu/modifyAdminMenu": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 메뉴 수정 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "act": {
                  "example": "any"
                },
                "menu_code": {
                  "example": "any"
                },
                "menu_name": {
                  "example": "any"
                },
                "menu_path": {
                  "example": "any"
                },
                "use_yn": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin_menu/setAdminMenuOrder": {
      "post": {
        "tags": [
          "관리자/Menu"
        ],
        "summary": "관리자 메뉴 관리 - 메뉴 순서변경 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "parentCode": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}